refactor(Calendar): Move CalendarDay to its own file with tests
