+++ packages/react/src/Calendar/CalendarBody.test.tsx
@@ -3,8 +3,6 @@
  * Copyright Gemeente Amsterdam
  */
 
-import type { AnchorHTMLAttributes } from 'react'
-
 import { render, screen } from '@testing-library/react'
 import { afterEach, describe, expect, it, vi } from 'vitest'
 
@@ -29,53 +27,6 @@ describe('CalendarBody', () => {
     render(<CalendarBody linkTemplate={(date) => `/day/${date.getDate()}`} locale="nl-NL" month={may2026} />)
 
     expect(screen.getAllByRole('link')).toHaveLength(31)
-  })
-
-  it('builds each link target with linkTemplate', () => {
-    render(<CalendarBody linkTemplate={(date) => `/day/${date.getDate()}`} locale="nl-NL" month={june2026} />)
-
-    expect(screen.getByRole('link', { name: 'maandag 15 juni 2026' })).toHaveAttribute('href', '/day/15')
-  })
-
-  it('gives each link an accessible name with the weekday and date', () => {
-    render(<CalendarBody linkTemplate={(date) => `/day/${date.getDate()}`} locale="nl-NL" month={june2026} />)
-
-    expect(screen.getByRole('link', { name: 'maandag 15 juni 2026' })).toBeInTheDocument()
-  })
-
-  it('renders a date as plain text when linkTemplate returns undefined', () => {
-    render(<CalendarBody linkTemplate={() => undefined} locale="nl-NL" month={june2026} />)
-
-    expect(screen.queryByRole('link')).not.toBeInTheDocument()
-    expect(screen.getByText('15')).toBeInTheDocument()
-  })
-
-  it('gives a plain-text date the same accessible weekday and date as a link', () => {
-    render(<CalendarBody linkTemplate={() => undefined} locale="nl-NL" month={june2026} />)
-
-    expect(screen.getByText('maandag 15 juni 2026')).toBeInTheDocument()
-  })
-
-  it('renders plain text when no linkTemplate is given', () => {
-    render(<CalendarBody locale="nl-NL" month={june2026} />)
-
-    expect(screen.queryByRole('link')).not.toBeInTheDocument()
-    expect(screen.getByText('15')).toBeInTheDocument()
-  })
-
-  it('renders dates with a custom link component', () => {
-    const CustomLink = (linkProps: AnchorHTMLAttributes<HTMLAnchorElement>) => <a {...linkProps} data-custom />
-
-    render(
-      <CalendarBody
-        linkComponent={CustomLink}
-        linkTemplate={(date) => `/day/${date.getDate()}`}
-        locale="nl-NL"
-        month={june2026}
-      />,
-    )
-
-    expect(screen.getByRole('link', { name: 'maandag 15 juni 2026' })).toHaveAttribute('data-custom')
   })
 
   it('marks today as the current date', () => {
