+++ packages/react/src/Calendar/CalendarBody.tsx
@@ -3,11 +3,9 @@
  * Copyright Gemeente Amsterdam
  */
 
-import type { AnchorHTMLAttributes } from 'react'
-
-import { clsx } from 'clsx'
-
 import type { CalendarProps } from './Calendar'
+
+import { CalendarDay } from './CalendarDay'
 
 export type CalendarBodyProps = {
   /** The month to display. */
@@ -16,46 +14,6 @@ export type CalendarBodyProps = {
 
 const isSameDay = (a: Date, b: Date) =>
   a.getFullYear() === b.getFullYear() && a.getMonth() === b.getMonth() && a.getDate() === b.getDate()
-
-const formatAccessibleDate = (date: Date, locale?: string) =>
-  new Intl.DateTimeFormat(locale, { day: 'numeric', month: 'long', weekday: 'long', year: 'numeric' }).format(date)
-
-const DefaultDateLink = (anchorProps: AnchorHTMLAttributes<HTMLAnchorElement>) => <a {...anchorProps} />
-
-type CalendarDayProps = {
-  readonly date: Date
-  readonly isCurrent: boolean
-} & Pick<CalendarBodyProps, 'linkComponent' | 'linkTemplate' | 'locale'>
-
-const CalendarDay = ({ date, isCurrent, linkComponent, linkTemplate, locale }: CalendarDayProps) => {
-  const DateLink = linkComponent ?? DefaultDateLink
-  const href = linkTemplate?.(date)
-  const ariaCurrent = isCurrent ? 'date' : undefined
-  const content = (
-    <>
-      <span aria-hidden={true}>{date.getDate()}</span>
-      <span className="ams-visually-hidden">{formatAccessibleDate(date, locale)}</span>
-    </>
-  )
-
-  if (href === undefined) {
-    return (
-      <span aria-current={ariaCurrent} className={clsx('ams-calendar__day', isCurrent && 'ams-calendar__day--current')}>
-        {content}
-      </span>
-    )
-  }
-
-  return (
-    <DateLink
-      aria-current={ariaCurrent}
-      className={clsx('ams-calendar__day-link', isCurrent && 'ams-calendar__day-link--current')}
-      href={href}
-    >
-      {content}
-    </DateLink>
-  )
-}
 
 export const CalendarBody = ({ linkComponent, linkTemplate, locale, month }: CalendarBodyProps) => {
   const year = month.getFullYear()
